create customhook ,transform the interfaces to file it #19
