+++ client/src/components/customHooks/UseRef.tsx
@@ -1,0 +1,23 @@
+import { useEffect, useRef } from 'react';
+
+const useRefFunction = (callback: Function) => {
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClick = (event: any) => {
+      if (ref.current && !ref.current.contains(event.target)) {
+        callback();
+      }
+    };
+
+    document.addEventListener('click', handleClick);
+
+    return () => {
+      document.removeEventListener('click', handleClick);
+    };
+  }, []);
+
+  return ref;
+};
+
+export default useRefFunction;
