+++ client/src/components/postCard/LikeItem.tsx
@@ -1,17 +1,7 @@
 import { NavLink } from 'react-router-dom';
+import ILike from '../../Interfaces/post/ILike';
 
-interface User{
-  id:number
-  name:string
-  avatar:string
-}
-interface ILike{
-  like: {
-    id:number
-    User:User
-  }
-}
-const LikeItem = ({ like: { id, User } } : ILike) => (
+const LikeItem = ({ like: { id, User } } :{ like:ILike}) => (
   <NavLink to={`/users/${id}`} className="likeItem">
     <img src={User.avatar} alt="" />
     <h4>{User.name}</h4>
